% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinPost.R
\name{betaBinPost}
\alias{betaBinPost}
\title{Beta binomial posteriors}
\usage{
betaBinPost(x, num.cores = 1L, tasks = 0L, verbose = TRUE, ...)
}
\arguments{
\item{x}{A numerical vector, a matrix or a data.frame object.}

\item{num.cores, tasks}{Parameters for parallel computation using
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}

\item{verbose}{if TRUE, prints the function log to stdout and a progress bar}

\item{...}{Further arguments for \code{\link{betaDistEstimation}} function.}
}
\description{
In a Bayesian framework, the parameters from Beta
distribution, \eqn{\alpha} and \eqn{\beta}, estimated from count data are
used as pseudo-counts to estimate the posterior probabilities:
\deqn{p = (\alpha + success)/(\alpha + \beta + trials)}.
}
\seealso{
\code{\link{betaDistEstimation}} and
\code{\link{estimateDirichDist}}
}
\author{
Robersy Sanchez <https://genomaths.com>
}
