% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDistEstimation.R
\name{betaDistEstimation}
\alias{betaDistEstimation}
\title{Nonlinear Parameter Estimation for Beta Distribution}
\usage{
betaDistEstimation(
  q,
  init.pars = c(1, 1),
  force.optim = FALSE,
  hessian = FALSE,
  method = "BFGS",
  gr = NULL,
  control = list(maxit = 500, abstol = (10^-8)),
  lower = -Inf,
  upper = Inf,
  seed = 123,
  ...
)
}
\arguments{
\item{q}{prior probabilities}

\item{init.pars}{initial parameter values. Defaults to alpha = 1 &
beta = 1, which imply the parsimony pseudo-counts greater than zero.}

\item{force.optim}{Whether to force the use of \code{\link[stats]{optim}}
function for the parameter estimation. Default is FALSE.}

\item{hessian}{if TRUE, the hessian of f at the minimum is returned.}

\item{method, control, lower, upper, gr}{(Optional). In the case that
\code{\link[stats]{nlm}} function fails, the methods, list of control
parameters, and bounders to be used with \code{\link[stats]{optim}} to
accomplish the parameter estimation.}

\item{control}{(Optional). In the case that \code{\link[stats]{nlm}}
function fails, a list of control parameters to be used with
\code{\link[stats]{optim}} function (see function help: ?optim) accomplish
the parameter estimation.}

\item{...}{Further parameter for \code{\link[stats]{nlm}} function.}
}
\value{
A list with components, which would vary depending on whether the
estimation was performed with \code{\link[stats]{nlm}} or
\code{\link[stats]{optim}}. In all the cases the list element carrying
the estimated parameters values is named \strong{\emph{parameters}}.
}
\description{
This function perform a rough estimation of the shape
    parameters of beta distribution
}
\details{
In order to obtain the estimates for shape parameters of beta
distribution, the squared of the difference between the empirical cumulative
distribution function (ecdf) & the theoretical cdf is minimized using the
Non-Linear Minimization function \code{\link[stats]{nlm}} 'stats' package.

If \code{\link[stats]{nlm}} function fails, then an estimation using
\code{\link[stats]{optim}} function is tried.
}
\examples{
### A random generation numerical values with Beta distribution
x1 <- rbeta(n = 1000, shape1 = 2, shape2 = 3)

### Parameter estimation with "nlm" function
betaDistEstimation(q = x1, gradtol = 1e-12, hessian = TRUE)

### Parameter estimation with "optim" function
betaDistEstimation(q = x1, force.optim = TRUE, hessian = TRUE)

}
\seealso{
\code{\link{betaBinPost}} and \code{\link{estimateDirichDist}}
}
\author{
Robersy Sanchez <https://genomaths.com>
}
