% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgeom.R
\name{bgeom}
\alias{bgeom}
\alias{dbgeom}
\alias{pbgeom}
\alias{qbgeom}
\alias{rbgeom}
\alias{bgeom2}
\alias{dbgeom2}
\alias{pbgeom2}
\alias{qbgeom2}
\alias{rbgeom2}
\title{Beta-Geometric}
\usage{
dbgeom(x, shape1, shape2, log = FALSE)

pbgeom(q, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qbgeom(p, shape1, shape2, qmax = 10000, lower.tail = TRUE, log.p = FALSE)

rbgeom(n, shape1, shape2)

dbgeom2(x, shape1, shape2, log = FALSE)

pbgeom2(q, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qbgeom2(p, shape1, shape2, qmax = 10000, lower.tail = TRUE, log.p = FALSE)

rbgeom2(n, shape1, shape2)
}
\arguments{
\item{x, q}{A numeric vector.}

\item{shape1, shape2}{shape parameters to pass to Beta distribution,
\code{\link[stats]{Beta}}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X<=x)}, otherwise, \eqn{P(X > x)}.}

\item{log.p, log}{logical; if TRUE, probabilities/densities p are returned
as log(p).}

\item{p}{A vector of probabilities.}

\item{qmax}{The maximum expected quantile. An argument to be used with
\emph{qbgeom} function.}

\item{n}{The number of observations}
}
\value{
The PDF, CDF, quantile and random generation functions for BGD.
}
\description{
Probability density function (PDF), cumulative density function (CDF),
quantile function, and random variate generation of the Beta-Geometric 
distribution (BGD). Herein, we consider the \strong{unshifted} sBGD PDF 
(dbgeom):

\deqn{f(x | \alpha,\beta) = 
\frac{B(\alpha + 1, \beta + x - 1)}{B(\alpha,\beta)}
}

for \eqn{x = 1,2,...}, and \eqn{0 < p \leq 1} and the \strong{shifted} 
PDF (dbgeom2):

\deqn{f(x | \alpha,\beta) = 
\frac{B(\alpha + 1, \beta + x)}{B(\alpha,\beta)}
}
with parameters \eqn{\alpha} (shape1) and \eqn{\beta} (shape1) as given in
references (1-2), for \eqn{x = 0,1,2,...}, and \eqn{0 < p \leq 1}, where 
\eqn{B(.)} is the \code{\link[base]{beta}}(a,b) function:

\deqn{B(a,b)= \frac{\Gamma(a)\Gamma(b)}{\Gamma(a+b)}}
}
\details{
Details about the BGD can be found in reference (1-3). 

## The Cumulative Distribution Function (CDF)
The CDF is computed as:

\deqn{F(x|\alpha, \beta) = \Sigma_{i=0}^{x} f(x |\alpha, \beta)}

## The quantile function

The quantile function is the inverse of the CDF, i.e., \eqn{F^{-1}}. Since
there is not analytical function for the CDF, we compute the quantile
\eqn{\hat{q}_i} for the probability value \eqn{p_i} as the argument
\eqn{x_i} that minimizes the difference 
\eqn{1/2 (F(x_i|\alpha, \beta) - p_i)}, formally:

\deqn{\hat{q}_i = \underset{x \in \mathbb{Z}^+}{\operatorname{Min}}
\left \{ \frac{1}{2} (F(x_i|\alpha, \beta) - p_i)^2 \:
| \: F(x_i|\alpha, \beta) \leq p_i \right \}}

## Random variate generation (random sampling)

To generate a random variate 
\eqn{X \sim \: F(x|\alpha, \beta)}, first, we
must generate a random variate \eqn{\pi} (probabilities): 
\eqn{\pi \sim \: B_x(x|\alpha, \beta)} (\code{\link[stats]{rbeta}}). Hence, 
we can generate \eqn{X \sim \: F(x|\alpha, \beta)} as:

\deqn{X = G^{-1}(R_B(n|\alpha, \beta) | \pi)}

where \eqn{R_B(n|\alpha, \beta)} stands for the function to generate random
variate from Beta distribution: \code{\link[stats]{rbeta}} and
\eqn{G^{-1}(p|\pi)} the quantile of the geometric distribution.
}
\examples{
## Generate random variates from the unshifted Beta-Geometric
set.seed(122)
x1 <- rbgeom(1e3, shape1 = 2, shape2 = 40)
summary(x1)

## The nonlinear fit 
cdf <- fit_cdf(x1, distNames = "Beta-Geometric I",
              plot = T)
cfs <- coefs(cdf)
cdf

## The histogram with the density (PDF) curve.
hist(x1, 600, freq = FALSE, xlim = c(0, 50),
     panel.first={points(0, 0, pch=16, cex=1e6, col="grey95")
         grid(col="white", lty = 1)})
x <- round(seq(min(x1), max(x1), by = 1))
lines(x, dbgeom(x, shape1 = 2, shape2 = 40), col = "red")
lines(x, dbgeom(x, shape1 = cfs[1], shape2 = cfs[2]), col = "dodgerblue")

## Kolmogorov-Smirnov tests can fail
set.seed(122)
x2 <- rbgeom(1e3, shape1 = 2, shape2 = 30)
summary(x2)

cdf <- fit_cdf(x2, distNames = "Beta-Geometric I",
              plot = TRUE)
cdf

## Bootstrap test for Goodness of fit (GoF)
## using Kolmogorov-Smirnov (for the sake of reducing compuational time,
## the sample.size was set small, default is NULL).
mcgoftest(x2, cdf, stat = "ks", sample.size = 100)


## Generate random variates from the shifted Beta-Geometric
set.seed(122)
x1 <- rbgeom2(1e3, shape1 = 2, shape2 = 40)
summary(x1)

## The nonlinear fit 
cdf_sf <- fit_cdf(x1, distNames = "Beta-Geometric II",
            plot = T)
cdf_sf

}
\references{
1. Weinberg, P. & Gladen, B.C. (1986). The Beta-Geometric distribution 
   applied to comparative fecundability studies. Biometrics, 42, 547-560.
2. Gupta, A. K., Nadarajah, S (2004). The Beta-Geometric Distribution. In 
   Handbook of Beta Distribution and Its Applications Statistics. 
3. Paul, Sudhir R. (2005) "Testing Goodness Of Fit Of The Geometric 
   Distribution: An Application To Human Fecundability Data,"
   Journal of Modern Applied Statistical Methods: Vol. 4 : Iss. 2 , 
   Article 8. DOI: 10.22237/jmasm/1130803620
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
