% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_crossval.R
\name{cdf_crossval}
\alias{cdf_crossval}
\title{Cross validation of Cumulative Distribution Function model}
\usage{
cdf_crossval(formula, pars, X, maxiter = 1024, ptol = 1e-12, minFactor = 1e-06)
}
\arguments{
\item{formula}{a nonlinear model formula including variables and parameters.
It will be coerced to a formula if necessary.}

\item{pars}{Estimated model parameters.}

\item{X}{Objective variable used to build the model.}

\item{maxiter, ptol}{Arguments for function
\code{\link[minpack.lm]{nlsLM}}}

\item{minFactor}{A positive numeric value specifying the minimum step-size
factor allowed on any step in the iteration. The increment is calculated
with a Gauss-Newton algorithm and successively halved until the residual sum
of squares has been decreased or until the step-size factor has been reduced
below this limit. Default value: 10^-6.}
}
\description{
This function returns A goodness-of-fit criteria for nonlinear
model selection, specifically, the cross-validation correlation coefficient R
(R.Cross.val).
}
\details{
The cross-validation correlation coefficient R (R.Cross.val) is
an estimator of the average cross-validation predictive power (1).
}
\references{
\enumerate{
    \item Stevens JP. Applied Multivariate Statistics for the Social
          Sciences. Fifth Edit. Routledge Academic; 2009.
 }
}
\seealso{
\code{\link{mcgoftest}} for Bootstrap test for Goodness of fit.
}
