% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDirichDist.R
\name{estimateDirichDist}
\alias{estimateDirichDist}
\title{Nonlinear Parameter Estimation for Dirichlet Distribution \eqn{}}
\usage{
estimateDirichDist(
  x,
  start = NULL,
  num.cores = 1L,
  tasks = 0L,
  seed = 123,
  refit = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix or a data.frame object carrying count data.}

\item{start}{Initial parameter values for \eqn{\lapha =
\alpha_1, ... , \alpha_n} (all positive reals). Defaults is NULL.}

\item{num.cores, tasks}{Parameters for parallel computation using
\code{\link[BiocParallel]{BiocParallel-package}}: the number of cores to
use, i.e. at most how many child processes will be run simultaneously
(see \code{\link[BiocParallel]{bplapply}} and the number of tasks per job
(only for Linux OS).}

\item{verbose}{if TRUE, prints the function log to stdout and a progress bar}

\item{...}{Further arguments for \code{\link{betaDistEstimation}} function.}
}
\value{
A vector of estimated parameter values
}
\description{
The parameter estimation is accomplished using a count data
matrix. The estimation is based on the fact that if a variable \eqn{x = (x_1,
x_2, ...x_n)} follows Dirichlet Distribution with parameters \eqn{\alpha =
\alpha_1, ... , \alpha_n} (all positive reals), in short, \eqn{x ~
Dir(\alpha)}, then \eqn{x_i ~ Beta(\alpha_i, \alpha_0 - \alpha_i)}, where
Beta(.) stands for the Beta distribution and \eqn{\alpha_0 = \sum \alpha_i}.

Dirichlet distribution is a family of continuous multivariate probability
distributions, a multivariate generalization of the Beta distribution.
}
\details{
As any non-linear fitting, results strongly depends on the start
parameter values.
}
\examples{
## This example requires for 'DirichletReg' package to installed.
\dontrun{
library(DirichletReg)
#' ## A random generation numerical vectors with
x <- rdirichlet(n = 1000, alpha = c(2.1, 3.1, 1.2))
head(x)

estimateDirichDist(x)
}


}
\seealso{
\code{\link{betaDistEstimation}} and \code{\link{betaBinPost}}
}
\author{
Robersy Sanchez <https://genomaths.com>
}
