% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCDF.R
\name{fitCDF}
\alias{fitCDF}
\title{Nonlinear fit of a commulative distribution function}
\usage{
fitCDF(varobj, distNames, plot = TRUE, plot.num = 1, distf = NULL,
  start = NULL, only.info = FALSE, maxiter = 10^4, maxfev = 1e+05,
  ptol = 1e-12, verbose = TRUE)
}
\arguments{
\item{varobj}{A a vector containing observations, the variable for which the
CDF parameters will be estimated.}

\item{distNames}{a vector of distribution numbers to select from the listed
below in details section, e.g. c(1:10, 15). If 'distNames' is not any of
current 20 named distributions, then it can be any arbitrary character
string, but the argument 'distf' must be given (see below).}

\item{plot}{Logic. Default TRUE. Whether to produce the plots for the best
fitted CDF.}

\item{plot.num}{The number of distributions to be plotted.}

\item{distf}{A symbol naming a cumulative distribution function (CDF) present
in the R session environment (Not a character string!). For example,
\strong{pnorm}, \strong{pgamma}, etc. If the function is not present in
the environment, then an error will be returned. It must given only if
'distNames' is not any of current 20 named distributions (see  details
below). Default is NULL.}

\item{start}{A named numerical vector giving the parameters to be optimized
    with initial values. This can be omitted for some of the named
    distributions (see Details). This argument will be used if provided for
    only one distribution. The default parameter values are:
\enumerate{
    \item norm = c( mean = MEAN, sd = SD )
    \item lnorm = c(meanlog = mean( log1p(X), na.rm = TRUE),
            sdlog = sd( log1p( X ), na.rm = TRUE))
    \item halfnorm = c(theta = sqrt(pi/2))
    \item gnorm = c( mean = MEAN, sigma = SD, beta = 2)
    \item tgnorm = c( mean = MEAN, sigma = SD, beta = 2)
    \item laplace = c( mean = MEAN, sigma = sqrt( VAR))
    \item gamma = c( shape = MEAN^2/VAR, rate = MEAN/VAR)
    \item gamma3p = c( shape = MEAN^2/VAR, rate = MEAN/VAR, mu = 0),
    \item ggamma4 = c(alpha = MEAN^2/VAR, scale = VAR/MEAN, mu = MIN,
                        psi = 1)
    \item ggamma3 = c( alpha = MEAN^2/VAR, scale = VAR/MEAN, psi = 1)
    \item weibull = c( shape = log( 2 ), scale = Q)
    \item weibull3p = c( mu = MIN, shape = log( 2 ), scale = Q)
    \item beta = c(shape1 = 1, shape2 = 2)
    \item beta3 = c(shape1 = 1, shape2 = 2, a = MIN)
    \item beta4 = c(shape1 = 2, shape2 = 3, a=0.9 * MIN, b=1.1 * MAX)
    \item bweibull = c(alpha=1, beta=2, shape = log( 2 ), scale = Q)
    \item gbeta = c( shape1 = 1, shape2 = 2, lambda = 1)
    \item rayleigh = c( sigma = SD )
    \item exp = c( rate = 1)
    \item exp2 = c( rate = 1, mu = 0)
}}

\item{only.info}{Logic. Default TRUE. If true, only information about the
parameter estimation is returned.}

\item{maxiter, maxfev, ptol}{Parameters to ontrol of various aspects of the
Levenberg-Marquardt algorithm through function
\code{\link[minpack.lm]{nls.lm.control}} from *minpack.lm* package.}

\item{verbose}{Logic. If TRUE, prints the function log to stdout}
}
\value{
After return the plots, a list with following values is provided:
    \itemize{
        \item aic: Akaike information creterion
        \item fit: list of results of fitted distribution, with parameter
            values
        \item bestfit: the best fitted distribution according to AIC
        \item fitted: fitted values from the best fit
        \item rstudent: studentized residuals
        \item residuals: residuals
    }
    After x = fitCDF( varobj, ...), attributes( x$bestfit ) yields:
    $names
    [1] "par" "hessian" "fvec" "info" "message" "diag" "niter" "rsstrace"
        "deviance"
    $class
    [1] "nls.lm"
    And fitting details can be retrived with summary(x$bestfit)
}
\description{
Usually the parameter estimation of a cummulative distribution
    function (*CDF*) are accomplished using the corresponding probability
    density function (*PDF*). DIfferent optimization algorithms can be used
    to accomplished this task and different algorithms can yield different
    esitmated parameters. Hence, why not try to fit the CDF directly?
}
\details{
The nonlinear fit (NLF) problem for CDFs is addressed with
    Levenberg-Marquardt algorithm implemented in function
    \code{\link[minpack.lm]{nls.lm}} from package *minpack.lm*. This function
    is inspired in a script for the function
    \code{\link[propagate]{fitDistr}} from the package propagate [1]. Some
    parts or script ideas from function \code{\link[propagate]{fitDistr}}
    are used, but here we to estimate CDF and not the PDF as in the case of
    "\code{\link[propagate]{fitDistr}}. A more informative are also
    incorporated. The studentized residuals are provided as well.
    The list (so far) of possible CDFs is:
    \enumerate{
        \item Normal \href{https://goo.gl/xaEAdT}{(Wikipedia)}
        \item Log-normal \href{https://goo.gl/a7MtYq}{(Wikipedia)}
        \item Half-normal \href{https://goo.gl/yxMF6T}{(Wikipedia)}. An
            Alternatively using a scaled precision (inverse of the variance)
            parametrization (to avoid issues if \eqn{\sigma} is near zero),
            obtained by setting \eqn{\theta=sqrt(\pi)/\sigma*sqrt(2)}.
        \item Generalized Normal \href{https://goo.gl/EPk8mH}{(Wikipedia)}
        \item T-Generalized Normal [2].
        \item Laplace \href{https://goo.gl/fCykV9}{(Wikipedia)}
        \item Gamma \href{https://goo.gl/cYkvar}{(Wikipedia)}
        \item 3P Gamma [3].
        \item Generalized 4P Gamma [3]
                \href{https://goo.gl/1n4kpW.}{(Wikipedia)}
        \item Generalized 3P Gamma [3].
        \item Weibull \href{https://goo.gl/WMXmQP}{(Wikipedia)}
        \item 3P Weibull \href{https://goo.gl/WMXmQP}{(Wikipedia)}
        \item Beta \href{https://goo.gl/893wzR}{(Wikipedia)}
        \item 3P Beta \href{https://goo.gl/893wzR}{(Wikipedia)}
        \item 4P Beta \href{https://goo.gl/893wzR}{(Wikipedia)}
        \item Beta-Weibull \href{https://goo.gl/dpaG8h}{ReliaWiki}
        \item Generalized Beta \href{https://goo.gl/UcVsct}{(Wikipedia)}
        \item Rayleigh \href{https://goo.gl/d9b3zv}{(Wikipedia)}
        \item Exponential \href{https://goo.gl/stVsi7}{(Wikipedia)}
        \item 2P Exponential \href{https://goo.gl/stVsi7}{(Wikipedia)}
    }
}
\examples{
set.seed(1230)
x1 <- rnorm(10000, mean = 0.5, sd = 1)
cdfp <- fitCDF(x1, distNames = "Normal", plot = FALSE)
summary(cdfp$bestfit)
}
\references{
\enumerate{
    \item Andrej-Nikolai Spiess (2014). propagate: Propagation of
          Uncertainty. R package version 1.0-4.
          http://CRAN.R-project.org/package=propagate
    \item Abramowitz, M. and Stegun, I. A. (1972) Handbook of Mathematical
          Functions. New York: Dover. Chapter 6: Gamma and Related Functions.
    \item Hand-book on STATISTICAL DISTRIBUTIONS for experimentalists
          (pag 73) by Christian Walck. Particle Physics Group Fysikum.
          University of Stockholm (e-mail: walck@physto.se).
  }
}
\seealso{
\code{\link[MASS]{fitdistr}} and \code{\link{fitMixDist}} and
    for goodness-of-fit: \code{\link{mcgoftest}}.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
