% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixDist.R
\name{fitMixDist}
\alias{fitMixDist}
\title{Nonlinear fit of Mixture distributions}
\usage{
fitMixDist(
  X,
  args = list(norm = c(mean = NA, sd = NA), weibull = c(shape = NA, scale = NA)),
  dens = TRUE,
  npoints = 100,
  kmean = FALSE,
  maxiter = 1024,
  prior = priorControl(),
  ftol = 1e-14,
  ptol = 1e-14,
  maxfev = 1e+05,
  equalPro = FALSE,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  usepoints,
  iter.max = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  seed = 123,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{numerical vector. It is user responsability to provide `X` values
that belong the definition domain of the functions from mixture
distribution.}

\item{dens}{Logic. Whether to use fit the 'PDF' or 'CDF'. Default is TRUE.}

\item{npoints}{number of points used in the fit of the density function or
NULL. These are used as histogram break points to estimate the empirical
density values. If \emph{npoints} = NULL and \emph{dens} = TRUE, then.
Kernel Density Estimation function \code{\link[stats]{density}} from
\emph{stats} package is used to estimate the empirical density. Default
value is 100.}

\item{kmean}{Logic. Whether to use \code{\link[stats]{kmeans}} algorithm to
perform the estimation in place of \code{\link[mclust]{Mclust}}. Deafult
is FALSE.}

\item{maxiter}{positive integer. Termination occurs when the number of
iterations reaches maxiter. Default value: 1024.}

\item{prior}{Same as in \code{\link[mclust]{Mclust}} function.}

\item{ftol}{non-negative numeric. Termination occurs when both the actual
and predicted relative reductions in the sum of squares are at most ftol.
Therefore, ftol measures the relative error desired in the sum of
squares. Default value: 1e-12}

\item{ptol}{non-negative numeric. Termination occurs when the relative error
between two consecutive iterates is at most ptol. Therefore, ptol
measures the relative error desired in the approximate solution.
Default value: 1e-12.}

\item{maxfev}{Integer; termination occurs when the number of calls to fn has
reached maxfev. Note that nls.lm sets the value of maxfev to
100*(length(par) + 1) if maxfev = integer(), where par is the list or
vector of parameters to be optimized.}

\item{equalPro}{An argument to pass to \code{\link[mclust]{emControl}}
function. Logical variable indicating whether or not the mixing
proportions are equal in the model. Default: equalPro = FALSE.}

\item{eps, tol}{Arguments to pass to \code{\link[mclust]{emControl}} function.}

\item{usepoints}{Integer. Computation by function
\code{\link[mclust]{Mclust}} could take long time when the sample size is
about >= 10000. This number can be used to extract a random sample of
size 'usepoints' and to do the estimation with it.}

\item{iter.max, nstart, algorithm}{Same as in \code{\link[stats]{kmeans}}.}

\item{seed}{Seed for random number generation.}

\item{verbose}{if TRUE, prints the function log to stdout and a progress bar}

\item{...}{Further arguments to pass to other functions like
\code{\link[mclust]{Mclust}} and \code{\link[stats]{density}}.}

\item{arg}{A list of named vectors with the corresponding named distribution
    parameters values. The names of the vector of parameters and the
    parameter names must correspond to defined functions. For example, if
    one of the involved distributions is the gamma density (see
    \code{\link[stats]{GammaDist}}), then the corresponding vector of
    parameters must be gamma = c(shape = 'some value', scale = 'some value').
    For the following distributions, the arguments can be provided with NULL
    values:
    \itemize{
        \item "norm" \href{https://goo.gl/xaEAdT}{(Wikipedia)}
        \item "halfnorm" \href{https://goo.gl/yxMF6T}{(Wikipedia)}.
        \item "lnorm" \href{https://is.gd/A5EJFi}{(Wikipedia)}
        \item "gnorm" \href{https://goo.gl/EPk8mH}{(Wikipedia)}
        \item "gamma" \href{https://goo.gl/cYkvar}{(Wikipedia)}
        \item "ggamma" \href{https://is.gd/oPHfX7}{(Wikipedia)}
        \item "beta" \href{https://goo.gl/893wzR}{(Wikipedia)}
        \item "beta" \href{https://goo.gl/893wzR}{(Wikipedia)}
        \item "laplace" \href{https://goo.gl/fCykV9}{(Wikipedia)}
        \item "weibull" \href{https://goo.gl/WMXmQP}{(Wikipedia)}
        \item "rayleigh" \href{https://goo.gl/d9b3zv}{(Wikipedia)}
        \item "exp" \href{https://goo.gl/stVsi7}{(Wikipedia)}
    }

Notice that the distribution given names correspond to the root-names as
given for R functions. For example, 'gamma' is the root-name for
functions \code{\link[stats]{GammaDist}}. See example, for more details. It
is necessary that any the root-name provided corresponds to a defined PDF
(if dens = TRUE) or to a defined CDF (if dens = FALSE) found in the session
`environment` (see ?environment).}

\item{fit.comp}{Logical. If FALSE, then starting parameter values for each
mixture component will be estimated suing function \code{\link{fitCDF}},
which could use more computational time. Default is FALSE}
}
\value{
A list with the model table with coefficients and goodness-of-fit
    results, the fitted model returned by function
    \code{\link[minpack.lm]{nls.lm}}, and a named list of fitted arguments.
}
\description{
This function performs the nonlinear fit of mixture
    distributions exploiting a firth approach on parameterized finite
    Gaussian mixture models obtained through the function
    \code{\link[mclust]{Mclust}} from package \emph{mclust}.
}
\details{
The approch tries to fit the proposed mixture distributions using a
    modification of Levenberg-Marquardt algorithm implemented in function
    \code{\link[minpack.lm]{nls.lm}} from \emph{minpack.lm} package that is
    used to perform the nonlinear fit. Cross-validations for the nonlinear
    regressions (R.Cross.val) are performed as described in reference [1]. In
    addition, Stein's formula for adjusted R squared (rho) was used as an
    estimator of the average cross-validation predictive power [1]. Notice
    that the parameter values must be given in way \emph{understandable}
    by the set of functions \code{\link{mixtdistr}} (see the example below).
    It is user responsability to provide `X` values that belong the
    definition domain of the functions from mixture distribution.
}
\examples{
#'
set.seed(1) # set a seed for random generation
## ========= A mixture of three distributions =========
phi = c(6/10, 4/10) #' Mixture proportions
## ---------------------------------------------------------

## === Named vector of the corresponding distribution function parameters
## must be provided
args <- list(gamma = c(shape = 2, scale = 0.1),
             weibull = c(shape = 3, scale = 0.5))
## ------------------------------------------------------------

## ===== Sampling from the specified mixture distribution ====
X <- rmixtdistr(n = 1e5, phi = phi , arg = args)
## ------------------------------------------------------------

## ===== Nonlinear fit of the specified mixture distribution ====
FIT <- fitMixDist(X, args = list(gamma = c(shape = NA, scale = NA),
                                 weibull = c(shape = NA, scale = NA)))

## === The graphics for the simulated dataset and the corresponding
## theoretical mixture distribution.
par(bg = "gray98",  mar = c(3, 4, 2, 1) )
hist(X, 90, freq = FALSE, las = 1, ylim = c(0, 5), xlim = c(0, 1),
     panel.first={points(0, 0, pch=16, cex=1e6, col="grey95")
       grid(col="white", lty = 1)},
     family = "serif", col = rgb(0, 0, 1, 0.),
     border = "deepskyblue", main = "Histogram of Mixture Distribution")
x1 <- seq(-4, 10, by = 0.001)
lines(x1, dmixtdistr(x1, phi = phi, arg = args), col = "red")
lines(x1, dmixtdistr(x1, phi = FIT$phi, arg = FIT$args), col = "blue")
legend(0.5, 4, legend = c("Theoretical Mixture PDF",
                          "Estimated Mixture PDF",
                          "Gamma distribution component",
                          "Weibull distribution component"),
       col=c("red", "blue", "magenta", "brown"), lty=1, cex=0.8)

## The standard definition of dgamma function (see ?dgamma)
lines(x1, dgamma(x1, shape = FIT$args$gamma[1],
                 scale = FIT$args$gamma[2]), col = "magenta")
## The standard definition of dgamma function (see ?dgamma)
lines(x1, dgamma(x1, shape = args$gamma[1],
                   scale = args$gamma[2]), col = "brown")

## The accuracy of the fitting depends on the the starting values
FIT <- fitMixDist(X, args = list(gamma = c(shape = 2.3, scale = 0.12),
                                 weibull = c(shape = 2.5, scale = 0.4)))
FIT$args
FIT$phi

}
\references{
1. Stevens JP. Applied Multivariate Statistics for the Social
    Sciences. Fifth Edit. Routledge Academic; 2009.
}
\seealso{
\code{\link[MASS]{fitdistr}}, \code{\link{fitCDF}},
    \code{\link{mixtdistr}}, and \code{\link{mcgoftest}}.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
