% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnorm.R
\name{hnorm}
\alias{hnorm}
\alias{dhnorm}
\alias{rhnorm}
\alias{phnorm}
\alias{qhnorm}
\alias{theta2sigma}
\alias{sigma2theta}
\title{Half-Normal distribution}
\usage{
dhnorm(x, theta = 1, log = FALSE)

phnorm(q, theta = 1, lower.tail = TRUE, log.p = FALSE)

qhnorm(p, theta = 1, sigma = NULL, lower.tail = TRUE, log.p = FALSE)

rhnorm(n, theta = 1)

theta2sigma(theta)

sigma2theta(sigma)
}
\arguments{
\item{theta}{numerical parameter, strictly positive (default 1).}

\item{q}{numeric vector}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X<=x],
otherwise, P[X > x]}

\item{log.p}{logical; if TRUE, probabilities/densities p are returned as
log(p).}

\item{sigma}{Standart deviation of the normal distribution. Here,
\eqn{\sigma = \sqrt \pi/(\theta\sqrt 2)}}

\item{n}{number of observations}
}
\value{
Half-normal PDF values (theta parameter) for dhnorm,
    Half-normal probability for phnorm, quantiles or Half-normal random
    generated values for rhnorm.
}
\description{
Probability density function (PDF), cummulative density function
    (CDF), quantile function and random generation for the Half-normal
    (hnorm) distribution.
}
\details{
An alternative parametrization to avoid issues when sigma is near
    zero is applied by using a scaled precision (inverse of the variance)
    obtained by setting \eqn{\theta=sqrt(\pi)/\sigma*sqrt(2)}. Details about
    these functions can be found in \href{https://goo.gl/yxMF6T}{Wikipedia}
    and in \href{https://is.gd/S9Kdss}{ MathWorld}. Notice that
    \eqn{\theta = 1} means \eqn{\sigma = \sqrt \pi/\sqrt 2}.
}
\examples{
set.seed(123) # set a seed
sigma = 1.2
theta = sigma2theta(sigma)
x <- rhnorm(n = 1e5, theta = theta)
hist(x, 100, freq = FALSE)
curve(dhnorm(x, theta = theta), col = "red", add = TRUE)

#' # Checking the function outputs for the logarithms of probabilities
x <- rhalfnorm(n = 10, theta = sigma2theta(2))
x1 <- phnorm(x, theta = sigma2theta(2), log = TRUE)
x2 <- phnorm(x, theta = sigma2theta(2), log = FALSE)
all(round(x1, 8) == round(log(x2), 8))

x3 <- dhnorm(x, theta = sigma2theta(2), log = TRUE)
x4 <- dhnorm(x, theta = sigma2theta(2), log = FALSE)
all(round(x3, 8) == round(log(x4), 8))
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
