% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_matrix.R
\name{hyp_matrix}
\alias{hyp_matrix}
\title{Compute the Hypothesis Matrix for MANOVA testing.}
\usage{
hyp_matrix(
  obj,
  par.mat = NULL,
  cov.mat = NULL,
  sample.size = NULL,
  groups = NULL
)
}
\arguments{
\item{obj}{A list of ANOVA R objects or linear/non-linear model objects.
If "obj" is not given, then the parmeters 'par.mat', 'cov.mat', and
'sample.size' must be given.}

\item{par.mat}{A matrix where each column contains the model parameters of
each curve.}

\item{cov.mat}{List of parameter variance-covariance matrices.}

\item{sample.size}{Vector with the sample sizes used to estimate each
models.}

\item{groups}{list of sets of indices that identify models from a same
group.}
}
\value{
A list carrying the following matrices and numerical data is
returned:

\itemize{
  \item "Pooled" sums of squares and cross product (SSCP) matrix: 'E'
  \item Degree of freedom for matrix E: 'df.e'
  \item The hypothesis SSCP matrix: 'H'
  \item Degree of freedom for matrix H: 'df.h'
  \item 'H + E': 'HpE.inv'
  \item Error covariance matrix: 'Se'
  \item Inverse of Se: Se.inv
  \item Sum of mean of Squares: 'W'
  \item Inverse of W: 'W.inv'
  \item Variance-covariance matrix from each model: 'Cov'
  \item Parameter matrix for the groups: 'par'
  \item Number of variables or model parameters: 'p'
  \item Total sample size: 'N'
  \item Number of models/groups: 'n'
  \item Sample size: 'm'
}
}
\description{
Given the summarized data provided, this function compute
the hypothesis matrices used in several multivaraite statistics.
}
\details{
To perform the comparison of a model under different parameter
values (different curve fits with the same model), the covariance matrices
and means from each set of model parameters can be used in place of the
covariance matrices and means for the variables inside of each group.
}
\examples{
## Generate a dataset
set.seed(1230)
x1 <- rnorm(1e3, mean = 2.5, sd = 1)
x2 <- rnorm(1e3, mean = 2.5, sd = 1)

line_1 <- 2.5 * x1 + 3 + runif(length(x1))/10
line_2 <- 2.5 * x2 + 3 + runif(length(x1))/10

## Fitting the linear models
lm_1 <- lm(line_1 ~ x1)
lm_2 <- lm(line_2 ~ x2)
res <- hyp_matrix(obj = list(model1 = lm_1, model2 = lm_2))
}
\seealso{
\code{\link{MANOVA}}.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
