% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtdistr.R
\name{mixtdistr}
\alias{mixtdistr}
\alias{dmixtdistr}
\alias{rmixtdistr}
\alias{pmixtdistr}
\alias{qmixtdistr}
\title{Mixture of Distribution Functions}
\usage{
dmixtdistr(x, pi, arg, log = FALSE, lower.tail = TRUE, log.p = FALSE)

pmixtdistr(q, pi, arg, lower.tail = TRUE, log.p = FALSE)

qmixtdistr(p, interval = c(0, 1000), pi, arg, lower.tail = TRUE,
  log.p = FALSE)

rmixtdistr(n, pi, arg)
}
\arguments{
\item{x, }{q vector of quantiles.}

\item{pi}{Numerical vector with mixture proportions, where (sum(pi) = 1).}

\item{arg}{A list of named vectors with the corresponding named distribution
parameters values. The names of the vector of parameters and the
parameter names must correspond to defined functions. For example, if
one of the involved distributions is the gamma density
(\code{\link[stats]{dgamma}}), then the corresponding vector of
parameters must be gamma = c(shape = 'some value', scale = 'some value').
See examples for more details.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x], 
otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}
}
\description{
Density, distribution function, quantile function and random 
    generation for mixture of distributions
}
\examples{
set.seed(123) # set a seed for random generation
# A mixture of three distributions
dist.fun <- c("gamma", "weibull", "lnorm")
pi = c(5/10, 3/10, 2/10) # Mixture proportions

# Named vector of the corresponding distribution function parameters
# must be provided
args <- list(gamma = c(shape = 20, scale = 1/10), 
            weibull = c(shape =  4, scale = 0.8),
            lnorm = c(meanlog = 1.2, sdlog = 0.08))
             
#  Sampling from the specified mixture distribution              
x <- rmixtdistr(n = 1e5, pi = pi , arg = args)

# The graphics for the simulated dataset and the corresponding theoretical 
# mixture distribution
hist(x, 100, freq = FALSE)
x1 <- seq(0, 10, by = 0.001)
lines(x1, dmixtdistr(x1, pi = pi, arg = args), col = "red")
}
