% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmR.R
\name{nlmR}
\alias{nlmR}
\title{Nonlinear Goodness-of-fit and Determination Coefficients}
\usage{
nlmR(object)
}
\arguments{
\item{object}{A 'nls' object, typically returned after performing a
nonlinear regression fit with function \code{\link[stats]{nls}} or
\code{\link[minpack.lm]{nls.lm}} or \code{\link[minpack.lm]{nlsLM}}.}
}
\description{
This function returns goodness-of-fit criteria for nonlinear
model selection.
}
\details{
Adjusted R squared (rho), the AIC, and BIC are estimated. The
Stein's formula for adjusted R squared (rho) is applied as an estimator of
the average cross-validation predictive power (1).
}
\references{
1. Stevens JP. Applied Multivariate Statistics for the Social
    Sciences. Fifth Edit. Routledge Academic; 2009.
}
\author{
Robersy Sanchez - 10/22/2020
}
