% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppCplot.R
\name{ppCplot}
\alias{ppCplot}
\title{P-P plot of Two-dimensional Copulas}
\usage{
ppCplot(X, Y, copula = NULL, margins = NULL, paramMargins = NULL,
  npoints = 100, method = "ml", smoothing = c("none", "beta",
  "checkerboard"), ties.method = "max",
  xlab = "Empirical probabilities", ylab = "Theoretical probabilities",
  glwd = 1.2, bgcol = "grey94", gcol = "white", dcol = "red",
  dlwd = 0.8, tck = NA, tcl = -0.3, xlwd = 0.8, ylwd = 0.8,
  xcol = "black", ycol = "black", cex.xtitle = 1.3,
  cex.ytitle = 1.3, padj = -1, hadj = 0.7, xcex = 1.3,
  ycex = 1.3, xline = 1.6, yline = 2.1, xfont = 3, yfont = 3,
  family = "serif", lty = 1, bty = "n", col = "black",
  xlim = c(0, 1), ylim = c(0, 1), pch = 20, las = 1, mar = c(4,
  4, 2, 1), font = 3, cex = 1, seed = 132, ...)
}
\arguments{
\item{X}{Numerical vector with the observations from the first margin
distribution.}

\item{Y}{Numerical vector with the observations from the second margin
distribution.}

\item{copula}{A copula object from class \code{\link[copula]{Mvdc}} or
string specifying all the name for a copula from package
\code{\link[copula]{copula-package}}.}

\item{margins}{A character vector specifying all the parametric marginal
distributions. See details below.}

\item{paramMargins}{A list whose each component is a list (or numeric
vectors) of named components, giving the parameter values of the marginal
distributions. See details below.}

\item{npoints}{Number of points used to build the P-P plot. The}

\item{method}{A character string specifying the estimation method to be used
to estimate the dependence parameter(s) (if the copula needs to be
estimated) see \code{\link[copula]{fitCopula}}.}

\item{smoothing}{character string specifying whether the empirical
distribution function (for F.n()) or copula (for C.n()) is computed (if
smoothing = "none"), or whether the empirical beta copula (smoothing =
"beta") or the empirical checkerboard copula (smoothing = "checkerboard")
is computed (see \code{\link[copula]{empCopula}}.}

\item{ties.method}{character string specifying how ranks should be computed
if there are ties in any of the coordinate samples of x; passed to
\code{\link[copula]{pobs}} (see \code{\link[copula]{empCopula}}.}

\item{xlab}{A label for the x axis, defaults to a description of x.}

\item{ylab}{A label for the y axis, defaults to a description of y.}

\item{glwd}{Grid line width.}

\item{bgcol}{Grid background color.}

\item{gcol}{Grid line color}

\item{dcol}{Diagonal line color.}

\item{dlwd}{Diagonal line color.}

\item{tck}{The length of tick marks as a fraction of the smaller of the width
or height of the plotting region. If tck >= 0.5 it is interpreted as a
fraction of the relevant side, so if tck = 1 grid lines are drawn. The
default setting (tck = NA) is to use tcl = -0.5.}

\item{tcl}{The length of tick marks as a fraction of the height of a line of
text. The default value is -0.5; setting tcl = NA sets tck = -0.01 which
is S' default.}

\item{xlwd}{X-axis line width.}

\item{ylwd}{Y-axis line width.}

\item{xcol}{X-axis line color.}

\item{ycol}{Y-axis line color.}

\item{cex.xtitle}{Cex for x-axis title.}

\item{cex.ytitle}{Cex for y-axis title.}

\item{padj}{adjustment for each tick label perpendicular to the reading
direction. For labels parallel to the axes, padj = 0 means right or top
alignment, and padj = 1 means left or bottom alignment. This can be a
vector given a value for each string, and will be recycled as necessary.}

\item{hadj}{adjustment (see par("adj")) for all labels parallel
(‘horizontal’) to the reading direction. If this is not a finite value,
the default is used (centring for strings parallel to the axis,
justification of the end nearest the axis otherwise).}

\item{xcex, ycex}{A numerical value giving the amount by which axis labels
should be magnified relative to the default.}

\item{xline, yline}{On which margin line of the plot the x & y labels must be
placed, starting at 0 counting outwards
(see \code{\link[graphics]{mtext}}).}

\item{xfont, yfont}{An integer which specifies which font to use for x & y
axes titles (see \code{\link[graphics]{par}}).}

\item{family, lty, bty, col, xlim, ylim, pch, las, mar, font}{Graphical parameters
(see \code{\link[graphics]{par}}).}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols should be magnified relative to the default. This starts as 1
when a device is opened, and is reset when the layout is changed, e.g.
by setting mfrow.}

\item{seed}{An integer used to set a 'seed' for random number generation.}

\item{...}{Other graphical parameters to pass to functions:
\code{\link[graphics]{abline}}, \code{\link[graphics]{mtext}} and
\code{\link[graphics]{axis}}.}
}
\value{
The P-P plot and invisible temporary object with the
    information to build the graphic which can be assigned to a variable
    to use in further plots or analyses.
}
\description{
The function build the P-P plot of Two-dimensional Copulas upon
    the knowledge of the margin distribution provided by the user. The
    empirical probabilities are computed using function
    \code{\link[copula]{empCopula}} from package
    \code{[copula-package]{copula}}.
}
\details{
Empirical and theoretical probabilities are estimated using the
    quantiles generated with the margin quantile functions. Nonlinear fit of
    margin distributions can be previously accomplished using any of the
    functions \code{\link{fitCDF}}, \code{\link[MASS]{fitdistr}}, or function
    \code{\link{fitMixDist}} for the case where the margins are mixture of
    distributions. \emph{npoints} random uniform and iid numbers from the
    interval [0, 1] are generated and used to evaluate the quantile margin
    distribution functions. Next, the quantiles are used to compute the
    empirical and theoretical copulas, which will be used to estimate the
    corresponding probabilities.
}
\examples{
set.seed(12)
margins = c("norm", "norm")
## Random variates from normal distributions
X <- rlnorm(200, meanlog =- 0.5, sdlog = 3.1)
Y <- rnorm(200, mean = 0, sd = 6)
cor(X,Y) ## Correlation between X and Y

parMargins = list( list(meanlog = 0.5, sdlog = 3.1),
                   list(mean = 0, sd = 10))

copula = "normalCopula"
npoints = 100

## The information to build the graphic is stored in object 'g'.
g <- ppCplot(X = X, Y = Y, copula = "normalCopula", margins = margins,
             paramMargins = parMargins, npoints = 20)

}
\seealso{
\code{\link{fitCDF}}, \code{\link[MASS]{fitdistr}},
    \code{\link{fitMixDist}}, and \code{\link{bicopulaGOF}}.
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
