% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CDFmodel.R
\name{print.CDFmodel}
\alias{print.CDFmodel}
\alias{print.CDFmodelList}
\title{Printing object from \emph{CDFmodel} class by simple print methods}
\usage{
\method{print}{CDFmodel}(x, digits = getOption("digits"), ...)

\method{print}{CDFmodelList}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \strong{\emph{CDFmodel}}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
An object from \emph{CDFmodel} classes is yielded by
function \emph{fitCDF}. This objects carries the information of a fitted
non-linear model.

An object from \emph{CDFmodelList} classes is yielded by
function \emph{fitCDF}. This objects carries the information of one or more
fitted non-linear models.
}
\details{
The definition of these class makes less complex the downstream
analyses.

The definition of these class makes less complex the downstream
analyses.
}
\keyword{internal}
