% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudents.R
\name{rstudents}
\alias{rstudents}
\alias{rstudents,missingORNULL-method}
\alias{rstudents,nls-method}
\title{Studentized Residuals from a Nonlinear Model}
\usage{
rstudents(model, varobj, method = c("Richardson", "simple", "complex"), ...)

\S4method{rstudents}{missingORNULL}(
  model,
  varobj,
  method = c("Richardson", "simple", "complex"),
  par,
  fun,
  residuals
)

\S4method{rstudents}{nls}(model, varobj, method = c("Richardson", "simple", "complex"))
}
\arguments{
\item{model}{An object from the ""nls" class. It can be derived. e.g., from
function \code{\link[stats]{nls}} or \code{\link[minpack.lm]{nlsLM}} from
the R packages \emph{stats} and \emph{minpack.lm}, respectively.}

\item{varobj}{Objective variable used build the \strong{model}.}

\item{method}{Argument for \code{\link[numDeriv]{grad}} function from
\emph{numDeriv} package. One of "Richardson", "simple", or "complex"
indicating the method to use for the approximation.}

\item{par}{\strong{model} parameters.}

\item{fun}{The named expression (not a character string) of the
fitted model. For example, the normal distribution function is given by
pnorm (see ?pnorm).}

\item{residuals}{Residuals from the model fitting.}
}
\value{
A vector of studentized residuals.
}
\description{
Given a fitted model and additional information, this function
computes the corresponding studentized residuals.
}
\details{
If the \strong{model} argument is provided, then only the
argument \strong{varobj} are required.
}
