% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\alias{se.nls}
\alias{se.nls.lm}
\alias{se.CDFmodel}
\alias{se.CDFmodelList}
\title{Sum of Squares of Error Statistics}
\usage{
se(model, stat)

\method{se}{nls}(model, stat = c("rmse", "sme"))

\method{se}{nls.lm}(model, stat = c("rmse", "sme"))

\method{se}{CDFmodel}(model, stat = c("rmse", "sme"))

\method{se}{CDFmodelList}(model, stat = c("rmse", "sme"))
}
\arguments{
\item{model}{An object from one of the classes: \code{\link[stats]{nls}},
\code{\link[minpack.lm]{nls.lm}}, CDFmodel or CDFmodelList, which are
obtained after running [fitCDF].}

\item{stat}{Select which statistic to return: 1) The root mean of the square
of error or the square errors of residuals.}
}
\value{
The requested statistic for each model.
}
\description{
Computes the residual sum of squares and the root of the mean of for squares
for models: \code{\link[stats]{nls}}, \code{\link[minpack.lm]{nls.lm}}, and
models CDFmodel and CDFmodelList obatained with [fitCDF].
}
\examples{
## Simulate data from Normal distribution
## and search for the best fitted model
x <-  rnorm(1000, mean = 5, sd = 2)
models <- fitCDF2(x, distNames = c(1, 7, 11, 12))

## Bases on the AIC criteria the best model is Weibull 3P
models

## In the current case Weibull is also the model with the lowest error
se(model)

}
