% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse}
\alias{sse}
\alias{sse.nls}
\alias{sse.nls.lm}
\alias{sse.CDFmodel}
\alias{sse.CDFmodelList}
\title{Sum of Squares of Error Statistics}
\usage{
sse(model, stat)

\method{sse}{nls}(model, stat = c("rmse", "sse"))

\method{sse}{nls.lm}(model, stat = c("rmse", "sse"))

\method{sse}{CDFmodel}(model, stat = c("rmse", "sse"))

\method{sse}{CDFmodelList}(model, stat = c("rmse", "sse"))
}
\arguments{
\item{model}{An object from one of the classes: \code{\link[stats]{nls}},
\code{\link[minpack.lm]{nls.lm}}, CDFmodel or CDFmodelList, which are
obtained after running [fitCDF].}

\item{stat}{Select which statistic to return: 1) The root mean of the square
of error or the square errors of residuals.}
}
\value{
The requested statistic for each model.
}
\description{
Computes the residual sum of squares and the root of the mean of for squares
for models: \code{\link[stats]{nls}}, \code{\link[minpack.lm]{nls.lm}}, and
models CDFmodel and CDFmodelList obatained with [fitCDF].

For a \code{\link[stats]{nls}} the RMSE is computed as:
\deqn{ y = resid(model)}
\deqn{RMSE = sqrt(mean(y^2))}
}
\examples{
## Simulate data from Normal distribution
## and search for the best fitted model
x <-  rnorm(1000, mean = 5, sd = 2)
models <- fitCDF2(x, distNames = c(1, 7, 11, 12))

## Bases on the AIC criteria the best model is Weibull 3P
models

## In the current case Weibull is also the model with the lowest error
sse(model)

}
