% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_crossval.R
\name{cdf_crossval}
\alias{cdf_crossval}
\alias{cdf_crossval,missingORNULL-method}
\alias{cdf_crossval,nls-method}
\alias{cdf_crossval,CDFmodel-method}
\alias{cdf_crossval,nls.lm-method}
\title{Cross validation of Cumulative Distribution Function model}
\usage{
cdf_crossval(model, ...)

\S4method{cdf_crossval}{missingORNULL}(
  model,
  formula,
  pars,
  q,
  logx = FALSE,
  min.val = NULL,
  loss.fun = c("linear", "huber", "smooth", "cauchy", "arctg"),
  maxiter = 1024,
  maxfev = 1e+05,
  ptol = 1e-12,
  minFactor = 1e-06
)

\S4method{cdf_crossval}{nls}(
  model,
  q,
  logx = FALSE,
  min.val = NULL,
  maxiter = 1024,
  ptol = 1e-12,
  minFactor = 1e-06
)

\S4method{cdf_crossval}{CDFmodel}(
  model,
  q,
  logx = FALSE,
  min.val = NULL,
  maxiter = 1024,
  ptol = 1e-12,
  minFactor = 1e-06
)

\S4method{cdf_crossval}{nls.lm}(
  model,
  formula,
  q,
  logx = FALSE,
  min.val = NULL,
  maxiter = 1024,
  ptol = 1e-12,
  minFactor = 1e-06
)
}
\arguments{
\item{formula}{No required for when a model from class
\strong{\emph{CDFmodel}} or \code{\link[stats]{nls}} is provided. Otherwise,
it will be nonlinear model formula including variables and parameters,
which will be coerced to a formula if necessary. For example, for a Gamma
model the formula will be: "Y ~ pgamma(q, shape, scale)", where
\code{\link[stats]{pgamma}} function is available in 'stats' R package.}

\item{pars}{Estimated model parameters.}

\item{q}{Objective variable used to build the model, typically called a
vector of quantiles. The model's formula must be expressed in terms of
variable \eqn{'q'}.}

\item{logx}{logical(1). If TRUE, then a logarithm transformation will be
applied: \eqn{q = log1p(q)}.}

\item{min.val}{A number denoting the lower bound of the domain where CDF
is defined. For example, for Weibull and GGamma \strong{\emph{min.val = 0}}.}

\item{loss.fun}{Described in \code{\link{fitCDF}}.}

\item{maxiter, ptol, maxfev}{Arguments for function
\code{\link[minpack.lm]{nlsLM}} and/or \code{\link[minpack.lm]{nls.lm}}.}

\item{minFactor}{A positive numeric value specifying the minimum step-size
factor allowed on any step in the iteration. The increment is calculated
with a Gauss-Newton algorithm and successively halved until the residual sum
of squares has been decreased or until the step-size factor has been reduced
below this limit. Default value: 10^-6.}
}
\description{
This function returns A goodness-of-fit criteria for nonlinear
model selection, specifically, the cross-validation correlation coefficient R
(R.Cross.val).
}
\details{
The cross-validation correlation coefficient R (R.Cross.val) is
an estimator of the average cross-validation predictive power (1).
}
\references{
\enumerate{
    \item Stevens JP. Applied Multivariate Statistics for the Social
          Sciences. Fifth Edit. Routledge Academic; 2009.
 }
}
\seealso{
\code{\link{mcgoftest}} for Bootstrap test for Goodness of fit.
}
