% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coefs}
\alias{coefs}
\alias{coefs,DirchModel-method}
\alias{coefs,BetaModel-method}
\alias{coefs,nls.lm-method}
\alias{coefs,CDFmodel-method}
\alias{coefs,CDFmodelList-method}
\title{Extract Model Coefficients}
\usage{
coefs(object, ...)

\S4method{coefs}{DirchModel}(object, output = c("alpha", "marginals"))

\S4method{coefs}{BetaModel}(object)

\S4method{coefs}{nls.lm}(object)

\S4method{coefs}{CDFmodel}(object)

\S4method{coefs}{CDFmodelList}(object)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
meaningful.}

\item{...}{Additional parameter not in use yet.}
}
\description{
This extend \coef{\link[stats]{coef}}
}
