% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCDF.R
\name{print.CDFmodel}
\alias{print.CDFmodel}
\title{Printing object from \emph{CDFmodel} class by simple print methods}
\usage{
\method{print}{CDFmodel}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object from class \strong{\emph{CDFmodel}}.}

\item{digits}{Number of significant digits to be used.}
}
\description{
An object from \emph{CDFmodel} classes is yielded by
function \emph{fitCDF}. This objects carries the information of a fitted
non-linear model.
}
\details{
The definition of these class makes less complex the downstream
analyses.
}
\keyword{internal}
