% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist.R
\name{unlist}
\alias{unlist}
\title{Flatten lists extended to any class}
\usage{
unlist(x, recursive = TRUE, use.names = TRUE)
}
\arguments{
\item{x}{Any list of R objects.}
}
\description{
Given a list 'x' of R objects from the same class and same
format, 'unlist' simplifies it to produce a new R object which contains all
the initial components which in 'x' object.
}
\details{

}
